blogpproject/requirements.txt
Django>=5.1.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
